reuse InfoModalContent
